New ABI version
Upstream changed their ABI version, so bump it in order to ensure
consistency with other distributions. There doesn't seem to have been an
actual change in ABI, so provide a compatibility package to make the
transition easier. Additionally, rename the -dev package - including the
ABI version in it implies that multiple versions can be installed in
parallel, which isn't actually the case.